certsva jadra presunuta do RSS
git-svn-id: https://zdrojaky.abclinuxu.cz/svn-abclinuxu/trunk@1692 be82d567-0701-4079-a524-1ea639e27e29

+++ src/cz/abclinuxu/servlets/html/edit/EditUser.java
@@ -681,7 +681,7 @@ public class EditUser implements AbcAction, Configurable {
         Persistence persistence = PersistenceFactory.getPersistance();
         for (Iterator iter = maintainedServers.iterator(); iter.hasNext();) {
             Integer id = (Integer) iter.next();
-            Server server = (Server) persistence.findById(new Server(id.intValue()));
+            Server server = (Server) persistence.findById(new Server(id));
             servers.add(server);
         }
         env.put(VAR_SERVERS, servers);
